Added more functions

+++ text_to_speech/text_to_speech.py
@@ -5,8 +5,15 @@ from playsound import playsound
 def play(file):
     playsound(file, True)
 
-def speak(text, lang = "en", slow = False):
-    file = "speech.mp3"
+def speak(text, lang = "en", slow = False, save = False, file = ""):
+    if save == True and file != "":
+        file = file
+    elif save == True and file == "":
+        file = "speech.mp3"
+    elif save == False:
+        file = "speech.mp3"
+    else:
+        raise ValueError(f"'{file}' is not defined")
     
     if slow == True:
         slow = True
@@ -16,7 +23,17 @@ def speak(text, lang = "en", slow = False):
         raise ValueError(f"'{slow}' is not defined")
     
     tts = gTTS(text=text, lang=lang, slow=slow)
-    tts.save(file)
 
-    play(file)
-    os.unlink(file)
+    if os.path.splitext(file)[1] == ".mp3":
+        tts.save(file)
+        
+        play(file)
+
+        if save == False:
+            os.unlink(file)
+        elif save == True:
+            pass
+        else:
+            raise ValueError(f"'{save}' is not defined")
+    else:
+        raise ValueError(f"'{file}' is not a valid mp3 format")
